ConditionalMovePacket can be used for PreCondNumMove

+++ RPG/src/rpgSource/moves/MagicMove.java
@@ -7,7 +7,7 @@ public class MagicMove extends PreConNumMove {
 	int mpUsage;
 
 	public MagicMove(int bd, String n, String d, Entities l, int mp) {
-		super(bd, n, d, l.getName() + " does not have enough mp to use " + n + ".", l, m -> m >= mp ? true : false);
+		super(bd, n, d, l.getName() + " does not have enough mp to use " + n + ".", l, m -> m.intValue() >= mp ? true : false);
 		mpUsage = mp;
 	}
